refactor code in users.db. update table definitions so prices are decimals

+++ db/init_db.js
@@ -63,7 +63,7 @@ async function createTables() {
                 id SERIAL PRIMARY KEY,
                 "userId" INTEGER,
                 FOREIGN KEY ("userId") REFERENCES users(id),
-                "totalPrice" INTEGER NOT NULL,
+                "totalPrice" DECIMAL NOT NULL,
                 "orderDate" DATE NOT NULL,
                 "orderStatus" status_type NOT NULL
             );
@@ -75,7 +75,7 @@ async function createTables() {
                 id SERIAL PRIMARY KEY,
                 title VARCHAR(255) UNIQUE NOT NULL,
                 description TEXT NOT NULL,
-                price INTEGER NOT NULL,
+                price DECIMAL NOT NULL,
                 quantity INTEGER NOT NULL,
                 active BOOLEAN DEFAULT true,
                 "categoryId" INTEGER NOT NULL,
@@ -87,8 +87,8 @@ async function createTables() {
                 "productId" INTEGER REFERENCES products(id) UNIQUE NOT NULL,
                 "orderId" INTEGER REFERENCES orders(id) NOT NULL,
                 quantity INTEGER NOT NULL,
-                "productPrice" INTEGER NOT NULL,
-                "totalPrice" INTEGER NOT NULL
+                "productPrice" DECIMAL NOT NULL,
+                "totalPrice" DECIMAL NOT NULL
             );
             CREATE TABLE reviews(
                 id SERIAL PRIMARY KEY,
